# Conflicts:
#	src/ui/lib/Signature.js

+++ src/ui/views/msgManage/warningConfig/index.vue
@@ -109,27 +109,32 @@ export default {
             }
         },
         delData(){
-            this.$confirm('此操作将永久删除该信息, 是否继续?', '提示', {
-				confirmButtonText: '确定',
-				cancelButtonText: '取消',
-				type: 'warning',
-			})
-            .then(() => {
-                request({
-                    url:`${this.$ip}/mms-warning/warningTemplate/delete/${this.selectId}`, 
-                    method: 'delete',
+            if(this.selectId!=null){
+                this.$confirm('此操作将永久删除该信息, 是否继续?', '提示', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning',
                 })
-                .then((data) => {
-                   this.$message({type: 'success',message: '删除成功'});
-                    this.getList();
+                .then(() => {
+                    request({
+                        url:`${this.$ip}/mms-warning/warningTemplate/delete/${this.selectId}`, 
+                        method: 'delete',
+                    })
+                    .then((data) => {
+                    this.$message({type: 'success',message: '删除成功'});
+                        this.getList();
+                         this.selectId = null;
+                    })
                 })
-            })
-            .catch(() => {
-                this.$message({
-                    type: 'info',
-                    message: '已取消删除',
+                .catch(() => {
+                    this.$message({
+                        type: 'info',
+                        message: '已取消删除',
+                    });
                 });
-            });
+            }else{
+                this.$message.error('请先选中一行数据');
+            }
         },
         getList(){
            request({
@@ -139,7 +144,6 @@ export default {
                 params:this.params
             })
             .then((data) => {
-               
                 this.tableData = extend({}, this.tableData, data.data);
             }).catch((error) => {
             
@@ -160,7 +164,6 @@ export default {
 <style scoped lang="scss">
 @import "@/ui/styles/common_list.scss"; 
 .sysParameter{
-    margin-top:40px;
     .main-content{
         /deep/ .mainTable{
             height: 600px;
